image_gen/requirements.txt
streamlit==1.39.0
torch==2.5.1
diffusers==0.31.0
pillow>=7.1.0,<11.0.0
transformers==4.46.1
accelerate==1.0.1
